Change Bluefruit library for recent Bluez versions
Recent (2018+?) versions of the Linux Bluez library/stack don't work with the official Adafruit lib anymore, and since the Bluefruit lib is officially deprecated, there won't be an update. I found that this version here works (Bluez 5.55 on Linux 5.10/Debian sid). 

I have not tested this on MacOS or anything.